Add missing validator and restructure Validators folder
