+++ backend/BookstoreWebAPI/Validators/Documents/CategoryValidator.cs
@@ -1,7 +1,7 @@
 ﻿using BookstoreWebAPI.Models.Documents;
 using FluentValidation;
 
-namespace BookstoreWebAPI.Validators
+namespace BookstoreWebAPI.Validators.Documents
 {
     public class CategoryValidator : AbstractValidator<CategoryDocument>
     {
