+++ backend/BookstoreWebAPI/Validators/FilterModels/CustomerFilterModelValidator.cs
@@ -1,0 +1,20 @@
+﻿using BookstoreWebAPI.Models.BindingModels.FilterModels;
+using FluentValidation;
+
+namespace BookstoreWebAPI.Validators.FilterModels
+{
+    public class CustomerFilterModelValidator : AbstractValidator<CustomerFilterModel>
+    {
+        public CustomerFilterModelValidator()
+        {
+            RuleFor(x => x.IsActiveString)
+                .Custom((isActiveString, context) =>
+                {
+                    if (!string.IsNullOrEmpty(isActiveString) && !bool.TryParse(isActiveString, out var val))
+                    {
+                        context.AddFailure("IsActive has to be true or false");
+                    }
+                });
+        }
+    }
+}
