+++ backend/BookstoreWebAPI/Validators/FilterModels/PromotionFilterModelValidator.cs
@@ -1,16 +1,16 @@
 ﻿using BookstoreWebAPI.Models.BindingModels.FilterModels;
 using FluentValidation;
 
-namespace BookstoreWebAPI.Validators
+namespace BookstoreWebAPI.Validators.FilterModels
 {
-    public class PromotionFilterModelValidator :AbstractValidator<PromotionFilterModel>
+    public class PromotionFilterModelValidator : AbstractValidator<PromotionFilterModel>
     {
         public PromotionFilterModelValidator()
         {
             RuleFor(x => x.IsOutdatedString)
                 .Custom((isOutdatedString, context) =>
                 {
-                    if (!string.IsNullOrEmpty(isOutdatedString) && !Boolean.TryParse(isOutdatedString, out var val))
+                    if (!string.IsNullOrEmpty(isOutdatedString) && !bool.TryParse(isOutdatedString, out var val))
                     {
                         context.AddFailure("IsOutdated has to be true or false");
                     }
