--- backend/BookstoreWebAPI/Validators/ProductFilterValidator.cs
@@ -1,12 +1,0 @@
-﻿using BookstoreWebAPI.Models.BindingModels.FilterModels;
-using FluentValidation;
-
-namespace BookstoreWebAPI.Validators
-{
-    public class ProductFilterValidator : AbstractValidator<ProductFilterModel>
-    {
-        public ProductFilterValidator()
-        {
-        }
-    }
-}
